python-app/requirements.txt
prometheus-client
flask[async]
flask-sock
requests
psutil
numpy
confluent-kafka
splunk-opentelemetry[all]